Create basic ROS entrainment node based on the demo
The speech entrainment ROS node will subscribe to messages to get an
incoming audio stream, to know whether or not someone is speaking in
that audio stream, to know if audio should be collected (i.e. it's the
participant's turn to speak), and what the name of the audio file to
morph is and the age of the participant. It will publihs log messages,
and send the morphed audio to the android-based robot over an rftp
stream rather than over ROS.